Added polling to see if the rabbitmq cluster is up
Because the broker could be deployed before the haproxy and rabbitmq
cluster, we added a polling attempts on the broker starup when trying to
initialize policies on existing vhosts. Because this can take some time,
polling has an timeout of ~ 30 mins.

The policy initializer still not connected to the cluter, we still need
to reorder the startup sequence.

Signed-off-by: Tiago Scolari <tiago@cloudcredo.com>