Extract rabbitmqctl wait into descriptive function
When this release is deployed the first time, the rabbitmq-clusterer
plugin will wait for all RabbitMQ nodes to come up before starting the
RabbitMQ Erlang app. The rabbitmqctl wait blocks until this app has
started. On the subsequent deploys it returns immediately but not on
the first deploy - we don't know why, need to ask Michael Klishin (the
author of rabbitmq-cluster) about the why. None of the rabbitmqctl user
commands will work until the RabbitMQ app is running.

In the new function, we are using a local variable to store the exit
code of the rabbitmqctl wait command. Mutating global variables is
really not a good idea.

Signed-off-by: Gerhard Lazu <glazu@pivotal.io>

[#119372579]