Skip admin access to all vhosts spec
This spec will need to be rewritten before we start refactoring the
rabbitmq-server.init ctl script. The current feedback loop for this spec
is 10-11 minutes which is just too long.

This spec is flaky because we can't know for sure when the RabbitMQ
users will end up being created. Because this runs synchronously, on all
nodes, after all nodes in the cluster are up and RabbitMQ application
has started, and also happens to loop over every vhost, we cannot be
certain when this will complete. A retry with a timeout is nothing more
than a band-aid which would make a bad situation worse.

[finishes #121842605]

Signed-off-by: Gerhard Lazu <glazu@pivotal.io>