Add RMQ checks as post-deploy script
These are meant to check that the RMQ cluster is healthy. We used to do
this in system tests, but they were slow and flaky which meant that we
didn't have the confidence that the cluster was configured correctly.
These checks are supposed to run automatically after every deployment.
To learn more about this feature see
http://bosh.io/docs/post-deploy.html

This is a follow-up from #121842605

[#124356151]

Signed-off-by: Gerhard Lazu <glazu@pivotal.io>