Run post-deploy script part of the rabbitmq-server init script
Our CF microbosh that is used by CI is an older version and does know of
post-deploy. This was introduced in v254, our BOSH director is 211.
Because the microbosh is shared by all teams, we do not want to perform
an upgrade and impact everyone. This is a temporary solution and should
be removed once BOSH post-deploy feature is GA.

[#124356151]

Signed-off-by: Zhou Yu <zyu@pivotal.io>