Fail specs fast in CI
We do not want to wait 1h+ when a spec fails. We do not want to watch
the job output for failed tests, fail fast and ensure the last job
output is the failed spec.

This is an experiment, only applicable to the master pipeline, feel free
to revert if it's not working out.

[finishes #126024625]

Signed-off-by: Gerhard Lazu <glazu@pivotal.io>