Do not log plugins & setup scripts
The redirects used with exec requires them to run under /bin/bash which
breaks the templating in setup.sh. They currently specify /bin/sh which
on Ubuntu stemcells is symlinked to /bin/dash. I don't know what exactly
/bin/sh is on CentOS stemcells, or whether setup.sh even works on
anything other than /bin/dash, but we know with certainty that it breaks
in /bin/bash.

For now, we are running these scripts as commands, not by sourcing them
in subshells, since the rabbitmq-server.init /bin/bash runtime will
overwrite any she-bang they specify.

We are also seeing that plugins.sh is actually failing because the
Erlang VM is not running, even though the plugins which it is trying to
enable are already enabled. Running `rabbitmqctl status` shows that all
3 apps/plugins are enabled and running: mqtt, stomp &
rabbitmq_management, in spite of these failures.

[#124356151]

Signed-off-by: Gerhard Lazu <glazu@pivotal.io>