Do not fail if the operator username is not defined
There is a check futher down which makes the operator username optional.
This flag will make the script fail if the operator username is not
defined.

Rather than depending on `set -u`, you might want to consider setting up
shellcheck integration with your code editor of choice. This feels more
like something that you would want to be caught via a linter (shellcheck
in this case) and not fail in a user unfriendly way.

[#126855183]

Signed-off-by: Gerhard Lazu <glazu@pivotal.io>