Remove the deployment after all tests pass
We want to ensure that after tests pass, we clean up the BOSH director
and delete the deployment. We do not want to update an existing
deployment and then run tests.

We are not deleting the deployment when the tests fail since we might
want to investigate/debug. This will require a manual delete deployment.

[finishes #131425441]

Signed-off-by: Zhou Yu <zyu@pivotal.io>