Do not run cluster check part of shutdown
Currently, cluster-check does not verify the health of the cluster. If
the RabbitMQ application is running, it considers the cluster to be
healthy, meaning that rabbitmq_clusterer plugin has started the RabbitMQ
application.

We might want to revisit this after we improve the cluster-check to
verify the cluster as a whole.

In a 3-node cluster with pause_minority, if 1 node is stopped, the
cluster-check will succeed. If 2 nodes are stopped, the cluster-check
will fail. We expect the cluster-check to fail in either case.

[#128761867]

Signed-off-by: Svett Ralchev <sralchev@pivotal.io>