Use the clusterer blob that we compiled
It turns out that the erlang plugin files can never be renamed, and for
this reason, the clusterer plugin generated by the rabbit core pipeline
cannot be used in the bosh release.

For now, we can work around this by compiling our own clusterer
plugin. In future we should help the core team change their pipeline to
produce plugins we can use.

[#131561809]

Signed-off-by: Ben Chapman <bchapman@pivotal.io>