Don't run upgrade preparation if the advanced op says not to
The core function of the upgrade check binary is to ensure that all
nodes in a cluster are running the same versions of erlang and
rabbitmq. As a part of this functionality, the upgrade check binary also
checks for netsplits. If a netsplit is in progress, it is impossible to
know if the nodes on the other side of the netsplit share our version or
not. For this reason, the upgrade check binary causes any deployment to
fail if there is a netsplit in progress.

IPSec can cause all the symptoms of a netsplit.

If we have a cluster of only a single node, this is fine, since there
can be no netsplit.

If we have a cluster of N nodes, but this is the first time any of those
nodes have been deployed, we already skip the upgrade check binary step,
as per story #131000637

If we have an existing cluster of N nodes, and we are upgrading it to a
cluster of N+2 nodes, then IPSec will cause all the symptoms of a
netsplit, and the upgrade check binary on the canary will cause the
deployment to fail. It is under these circumstances that an advanced
operator may wish to disable this check.

[finishes #132317847]

Signed-off-by: Gareth Smith <gsmith@pivotal.io>