Extract prepare_for_upgrade into pre-start
Monit start/stop/restart for a job will timeout in 30 seconds by
default. Stopping other nodes for breaking upgrades in a used cluster
will definitely take more than 30 seconds. We are moving this concern
into pre-start.

We want to simplify rabbitmq-server.init even further, this is a step in
that direction.

[#131000981]

Signed-off-by: Gerhard Lazu <glazu@pivotal.io>