Skip failing test
This test is currently neither in our main testing workflows nor it has
been refactored to work with manifest v2. Shall we keep it? More details
in the chore.

In the meanwhile, I am commenting it out because it is creating a lot of
noise when doing `bundle exec rspec`.

[#142614283]