Prepare merge from master
The master pipeline have now two different categories of tests:
integraiton and system.

In the integraiton set we have only tests related to the rabbitmq bosh
release itself. Nothing related to the multitenant broker is in that set
anymore. The integration tests will be executed by the pipeline thanks
to the `scripts/integration-tests` script. For now this script does
nothing. Once the master branch gets merged here this script is going to
execute the rabbitmq integration tests.

In the system set we have more broad tests, like tests that need the
multitenant broker deployed alongside with the rabbitmq release.

[#143407365]

Signed-off-by: Zhou Yu <zyu@pivotal.io>