Add property to toggle native cluster formation
Although, at this point the variable is hardcoded to true since we only
support native clustering formation. This means that the property
`rabbitmq-server.use_native_clustering_formation` is ignored.

This is part of our strategy to introduce small changes without breaking
anything in order to be always in a releasable state.

[#144893525]

Signed-off-by: Abhishek Chander <achander@pivotal.io>