Add toggle to activate rabbitmq_clusterer plugin
In order to activate the clusterer plugin you need to set the
`rabbitmq-server.use_native_clustering_formation = true`. Right now,
this is the default behaviour, which means that if the property is not
provided the native clustering formation will be used.

However, there's currently a problem with this design: if the property
is `rabbitmq-server.use_native_clustering_formation = true` and the
`rabbitmq_clusterer` is present in `rabbitmq-server.plugins` list, the
plugin will be enabled without the required configuration.

In order to use the rabbitmq_clusterer plugin just set
`rabbitmq-server.use_native_clustering_formation = false`.

[#144893525]

Signed-off-by: Abhishek Chander <achander@pivotal.io>