Embedd cluster tests at the release level
Currently we have cluster tests at the system level. This means the
tests need CF + BOSH + multitenant-broker in order to work. This is not
ideal for two reasons:

1. the feedback loop gets longer
2. the tests tend to fail more ofter, because many components are involved

To address these points, we're experimenting with what we called "embedded
release tests". Embedded release tests are jobs deployed colocated with
the jobs they are testing. This tests run checks during the `post-deploy` phase.
Beyond the 2 points mentionned above, we have the advantage of being
able to access all the tools, files, etc. available inside the
deployment, like `rabbitmqctl`, etc.

These test jobs are meant to be used during the development phase
(locally, within the pipeline). They are not meant to be used by
operators.

[#147032231]

Signed-off-by: Denise Yu <dyu@pivotal.io>
Signed-off-by: Hannes Hoerl <hhoerl@pivotal.io>