Resolv DNS names to IPs
Why do we need to resolve the DNS names to IPs?
In erl_netrc we cannot alias an DNS name to our md5-hashed hostnames and BOSH
does not guarantee that it returns an IP for a link's address. Thus, to be
safe we just convert all DNS names to IPs.

[#148653655]

Signed-off-by: Hannes Hoerl <hhoerl@pivotal.io>