Remove flaky test due to a race in bosh
* When the Erlang cookie changes in a deployment, any additional changes
may not be applied to all nodes.

This is because the Canary node will shut down all other nodes in the
cluster, but they are also monitored by `monit`, which will notice the
RMQ process is no longer running and restart it.

If the configuration changes have _not_ been applied at this point, the
server will restart with the old configuration, which is not ideal. We
have seen a number of build failures where this is the root cause, and
it can take a number of forms (e.g. the node still has an old value for
the FD limit, or has an old username/password), or (some of) the plugins
aren't loaded.

Another possible outcome is that the `post-deploy` task fails. This
should be fixable by running the deployment again.

There's a bug to be played on #149802915 and we should bring this test
back

[#152408792]