Do not generate the cookie in startup scripts.
The way the cookie was generated was not secure (it was predictable).
The scripts no longer generate the cookie - you should either specify it
explicitely or use a BOSH variable to generate it.

Cookie change requires cluster shutdown. You may specify your existing
cookie value to avoid downtime (but you should still rotate the cookie
at some point if it was generated by the script).

[#163054709]

Co-authored-by: Michal Kuratczyk <mkuratczyk@pivotal.io>