Move 'load_definitions' from SERVER_START_ARGS to rabbitmq.conf file
The command
`SERVER_START_ARGS='-rabbitmq_management load_definitions "/path/to/def.json"' rabbitmq-server`
worked with RabbitMQ 3.8.3 but fails with 3.8.4

Therefore, we use rabbitmq.conf to point to the definitions.json.

Due to historic reasons and in order to not break the
cf-rabbitmq-release API (i.e. to not change the BOSH job properties)
this release will render both the classic config (rabbitmq.config) and
the new config (rabbitmq.conf). The classic config - i.e. when the
rabbitmq-server.config  property is set - will take precedence over the
new rabbitmq.conf.

[#173134306]

Co-authored-by: David Ansari <dansari@pivotal.io>