Make cluster_name configurable
Before this commit, we set the cluster_name to be the BOSH deployment name.
This made it easy to filter in Prometheus based on cluster_name (i.e. BOSH deployment name).

However, we observed in production that this change broke the usage of
the cluster_name in the rabbitmq_federation plugin. In particular due to
a cluster_name change, it duplicated the number of federated queues.

Therefore, only set cluster_name if explicitly requested.

[#173565338]